@@ -1,11 +1,11 @@
 name: osc
 description: A simple Open Sound Control (OSC) protocol implementation for Dart.
-version: 0.0.4
+version: 1.0.0-dev
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dev_dependencies:
-  ansicolor: ^1.0.0
+  ansicolor: ^2.0.0-nullsafety.0
   pedantic: ^1.0.0
-  test: ^1.5.0
+  test: ^1.16.0-nullsafety
